+++ provisioner/worker/plugins/providers/fog_provider/lib/fog_provider/joyent.rb
@@ -77,6 +77,7 @@ class FogProviderJoyent < Provider
       bootstrap_ip = ip_address(server)
       if bootstrap_ip.nil?
         log.error 'No IP address available for bootstrapping.'
+        raise 'No IP address available for bootstrapping.'
       else
         log.debug "Bootstrap IP address #{bootstrap_ip}"
       end
@@ -90,8 +91,17 @@ class FogProviderJoyent < Provider
       set_credentials(@task['config']['ssh-auth'])
       # Validate connectivity
       Net::SSH.start(@result['result']['ipaddress'], @task['config']['ssh-auth']['user'], @credentials) do |ssh|
-        ssh_exec!(ssh, 'ping -c1 www.opscode.com', 'Validating external connectivity and DNS resolution via ping')
-        ssh_exec!(ssh, "hostname #{@task['config']['hostname']}", 'Temporarily setting hostname')
+        # Backwards-compatibility... ssh_exec! takes 2 arguments prior to 0.9.8
+        ssho = method(:ssh_exec!)
+        if ssho.arity == 2
+          log.debug 'Validating external connectivity and DNS resolution via ping'
+          ssh_exec!(ssh, 'ping -c1 www.opscode.com')
+          log.debug 'Temporarily setting hostname'
+          ssh_exec!(ssh, "hostname #{@task['config']['hostname']}")
+        else
+          ssh_exec!(ssh, 'ping -c1 www.opscode.com', 'Validating external connectivity and DNS resolution via ping')
+          ssh_exec!(ssh, "hostname #{@task['config']['hostname']}", 'Temporarily setting hostname')
+        end
       end
       # Return 0
       @result['status'] = 0